model and media files update
Currently updated final model and part STEP files for the following models. Some of the elements inside the model are definitely subject to change because the body is designed to maximize the print bed volume. I am experimenting with the fitment of all the pieces and the accuracy of the prints. Electrical and software work will begin in the coming new year due to exams.

Media folder contains some renders with all the components places inside!